feat: Out-of-tree cargo workspaces
This patch adds handling for out-of-tree workspaces.  The feature is
supported by cargo and requires that the root workspaces declare its
members in the key
```
[workspace]
members = [
   <RELATIVE_PATH_TO_MEMBERS>
]
```
and the members point back to the root workspace with the
following declaration:
```
[package]
workspace = <RELATIVE_PATH_TO_ROOT>
```

This patch complements the discovery by handling the members (in- or
out-of-tree) explicitly if the cargo manifests features the entries
mentioned above.

Closes #3450